Big refactor. Moved markers, joins, traces and vectors to separate modules. No logic has changed
